Refactor to get rid of endorse-related session variables.
The endorse part is simple. boxMsg2.php accepts the endorse info
as URL parameters instead of session variables and then includes
them in the form. The JS on that page can then determine when to
trigger the AJAX endorsing script and pass the necessary parameters
as part of the request.

In the process of re-writing all instances of endorsing, I made some
additions to the TenderModule class to make the SimpleTender parser
unnecessary. Rather than hardcoding a few tender values in that parser
for special handling the TenderModule itself can decide what to do
if no amount was specified.